+++ models/temperature.js
@@ -6,7 +6,7 @@ var Temperature = sequelize.define('Temperature', {
 
   id: Sequelize.INTEGER,
 
-  data: Sequelize.DOUBLE(8, 2),
+  data: Sequelize.FLOAT(8, 2),
 
   user_id: Sequelize.INTEGER,
